If revision selected requires JavaScript templates return them with response and add to page.
The id of the JS templates like "tmpl-*" are now saved in revview.js_templates. They're sent with the request.
The HTML originated from triggering wp_footer action, is parsed and JS templates are extracted. If their id is not in the list of the ids of templates already loaded, an array with the template HTML id attribute, and the JS template content is returned.
The id is then added to revview.js_templates.