Handle loading latest revision (current content) when UI starts and preserve latest revision viewed when user closes the UI.
Next time user opens UI (if a reload hasn't ocurred) the revision loaded will be the last one he was viewing. This was handled before with the difference that previously we were saving the latest revision HTML in the page as jQuery objects and restoring them afterwards.
Now the process is no different than viewing another revision, which makes it consistent to trigger scripts, load styles and js templates.