Suppress HTML entities notices (and other warnings and notices).
The JS templates string is loaded as XML to be parsed because in HTML mode it's invalid to place a <p> inside a <script>. However, in this mode there are unrecognized entities such as &times; and it throws notices For this reason this warning is suppressed. It's ok since we won't be using the result of the loading as is but instead it will first be parsed to find script tags.