Implement scratch area using an iframe.
The revision requested is first rendered in the iframe, scripts executed and then the relevant nodes for title, excerpt and content are extracted and placed into their corresponding locations in the top frame.
The admin bar is not rendered in the iframe.
The iframe location has the query string ?revview=render to identify it.