Implement revision viewer loading the revision in a iframe, leaving the revision selection UI in the top frame.
After the page is reloaded with a query string to indicate we're into a Revview page, the template is intercepted and replaced with a template that contains only the revision selection UI and the iframe where the revisions are loaded.
When a revision is requested, the title, content and excerpt extracted from REST API are sent and applied using the filters the_title, the_content and the_excerpt.