Go back to add the View Revisions button with JS. Apply fix for themes with body with position other than static.
Reason for this is that some themes doesn't seem to be clearing the $wp_query variable after custom loops and the permalink returned by get_permalink() was incorrect in views like a page view with a custom loop inside.

A different approach would have been to save the queried object ID very early and then use it, but to avoid passing an ID around and getting this to be unnecessarily complex and since there would already be a JS file loaded in front end to fix the body with positioning other than static, it was done with JS since it's very straightforward.