When fetching a revision, send only requested revision id and load its properties.
This is so we don't send title, content and excerpt from front end to compose the post. A case where this wouldn't be desirable (a bit extreme but not impossible) would be a user using a shortcode that renders content, for example, an accordion. The shortcode has an identifier parameter so it knows which accordion with content to load. A user used this shortcode and wrapped it in a shortcode provided by the theme to make it available only to logged in users:

[logged_in_content]
    [accordion id="1234"]
[/logged_in_content]

An attacker could send the same shortcode with the identifier and without the encompassing shortcode that makes it private:

[accordion id="1234"]

and thus obtain the content that was supposed to be hidden.

To avoid this, we only send the revision id, check that its parent is public, and get the revision's unaltered title, content and excerpt from db.