Go to original permalink instead of going back in browser history.
A visitor might have been given the direct link to the revision selection UI and open it in a new tab. In this case they won't have a previous history to go back and will end up in a blank/start tab. This change will ensure that when the revision selection UI is closed, the visitor will be taken to the current (and latest revision) entry.