Don't show admin bar on top frame when viewing revisions but on iframe.
The reason is that showing admin bar in top frame will leave it unsynced in relationship with content displayed. For example, Query Monitor plugin information might not reflect the revision loaded in iframe.
A better alternative is to add the admin bar in the revision iframe.

This is still debatable, because the admin bar in this context is a bit redundant since if user wants to do some task he can go to to the main entry.